Refactor: React query 적용
